moved bookreview, characters, and charactergroups to subdirectory
