+++ src/app/app.module.ts
@@ -5,7 +5,7 @@ import { HttpClientModule } from '@angular/common/http';
 import { RouterModule, Routes } from '@angular/router';
 
 import { AppComponent } from './app.component';
-import { BooksComponent } from './books/books.component';
+import { BookReviewComponent } from './books/book-review/books.component'
 import { CloneWarsReviewComponent } from './clone-wars-review/clone-wars-review.component';
 import { CloneWarsChronologicalComponent } from './clone-wars-chronological/clone-wars-chronological.component';
 import { TableComponent } from './utils/table/table.component';
@@ -13,8 +13,8 @@ import { TableComponent } from './utils/table/table.component';
 import { TableModule } from 'primeng/table';
 import { TableDataClassPipe } from './pipes/table-data-class.pipe';
 import { BookSectionComponent } from './utils/book-section/book-section.component';
-import { CharactersComponent } from './characters/characters/characters.component';
-import { CharacterGroupComponent } from './characters/character-group/character-group.component';
+import { CharactersComponent } from './books/characters/characters.component';
+import { CharacterGroupComponent } from './books/character-group/character-group.component';
 
 const routes: Routes = [
   { path: '', component: AppComponent }
@@ -23,7 +23,7 @@ const routes: Routes = [
 @NgModule({
   declarations: [
     AppComponent,
-    BooksComponent,
+    BookReviewComponent,
     CloneWarsReviewComponent,
     CloneWarsChronologicalComponent,
     TableComponent,
