+++ src/app/books/book-review/books.component.scss
@@ -1,4 +1,4 @@
-@import "../../styles.scss";
+@import "../../../styles.scss";
 
 .book-note {
   margin-bottom: 17px;
