+++ src/app/books/book-review/books.component.ts
@@ -1,14 +1,14 @@
 import { Component, OnInit } from '@angular/core';
-import { ReviewClass } from '../interfaces/enums';
-import { Book } from '../interfaces/interfaces';
-import { BookService } from '../services/book.service';
+import { ReviewClass } from '../../interfaces/enums';
+import { Book } from '../../interfaces/interfaces';
+import { BookService } from '../../services/book.service';
 
 @Component({
   selector: 'app-books',
   templateUrl: './books.component.html',
   styleUrls: ['./books.component.scss']
 })
-export class BooksComponent implements OnInit {
+export class BookReviewComponent implements OnInit {
 
   public bestCanonBooks: Book[];
   public goodCanonBooks: Book[];
