+++ src/app/books/character-group/character-group.component.ts
